+++ backend/scripts/seed-db.ts
@@ -4,10 +4,19 @@ const prisma = new PrismaClient();
 
 async function seedDatabase() {
 	try {
-		const projectData = [{ color: "#444", name: "Other" }];
-
-		await prisma.project.createMany({ data: projectData });
+		const project = await prisma.project.create({
+			data: { color: "#444", name: "Other" }
+		});
 		console.log("✅ Projects seeded successfully!");
+
+		await prisma.section.create({
+			data: {
+				project_id: project.id,
+				name: "Other"
+			}
+		});
+
+		console.log("✅ Sections seeded successfully!");
 
 		console.log("✅ Database seeded successfully!");
 	} catch (error) {